server/requirements.txt
bcrypt==4.2.0
fastapi==0.115.0
pydantic==2.9.2
annotated-types==0.7.0
pydantic_core==2.23.4
typing_extensions==4.12.2
starlette==0.38.5
anyio==4.6.0
idna==3.10
sniffio==1.3.1
mysql-connector-python==9.0.0
pipdeptree==2.23.4
packaging==24.1
pip==24.2
python-dotenv==1.0.1
uvicorn==0.30.6
click==8.1.7
colorama==0.4.6
h11==0.14.0
fastapi[standard]
fastapi-pagination
pysqlite3